Update README.md
If you downloaded the first library that I uploaded a week ago, the typedef of the destroy function to pass to the hashtable destructor, is changed, it's argument is not a pointer to mValue union but a void*. After a bit of reasoning, we can safely pass the ->v member and not more the union.
The other change is the function read_utf8_seq() in the early version it required three argument, the third was the length of the stream,  also here with a bit of reasoning we can safely assume that it's not required . The challenge here, is to understand how many characters to skip in the case of a bad sequence. We can found it out without the length.